requirements.txt
colorthief
openrgb-python
watchdog